+++ VsIntegration/Spect.Net.CodeDiscover.ProjectType/CodeDiscoverProjectTreePropertiesProvider.cs
@@ -52,18 +52,9 @@ namespace ZXSpectrumCodeDiscover
                 case "Z80Test":
                     propertyValues.Icon = ImageMonikers.Z80TestIconImageMoniker.ToProjectSystemType();
                     break;
-                // --- Temporarily removed - I need nicer folder icons
-                    //case "Folder":
-                    //    if (propertyContext.ItemName == "TapeFiles")
-                    //    {
-                    //        propertyValues.Icon = ImageMonikers.TapeFolderImageMoniker.ToProjectSystemType();
-                    //        propertyValues.ExpandedIcon = ImageMonikers.TapeFolderImageMoniker.ToProjectSystemType();
-                    //    } else if (propertyContext.ItemName == "Z80CodeFiles")
-                    //    {
-                    //        propertyValues.Icon = ImageMonikers.Z80FolderImageMoniker.ToProjectSystemType();
-                    //        propertyValues.ExpandedIcon = ImageMonikers.Z80FolderImageMoniker.ToProjectSystemType();
-                    //    }
-                    //    break;
+                case "Vfdd":
+                    propertyValues.Icon = ImageMonikers.FloppyIconImageMoniker.ToProjectSystemType();
+                    break;
             }
         }
     }