+++ CHANGELOG.md
@@ -6,6 +6,19 @@ of Spectrum 128K and Spectrum +3, (Bit 5=1, of the `$7FFD` port), and the ALLRAM
 __FETAURE__: While the emulator runs, the memory bank status (ZX Spectrum Memory and Z80 Disassembly tools) is continuously
 updated.
 
+### Version 1.9.0
+
+__FEATURE:__ Floppy drive emulation added to Spectrum +3e.
+* Drive A: and drive B: supported (0, 1, or 2 floppy configurations)
+* You can choose the configuration when creating a new project
+* All four floppy disk types supported by Spectrum +3 (Spectrum +3, CPC Data, CPC System, Amstrad PCW)
+can be created, inserted, or ejected.
+* Disks can be marked as write protected
+* The contents of the disks are stored in `.vfdd` files that are now part of
+the ZX Spectrum project
+* `.vfdd` files have an editor that displays the format information of the virtual disk
+
+
 ### Version 1.8.0
 
 __FEATURE__: Macros are now supported.